Moved ParseException to proper package
